+++ springql-core/examples/doc_app1.rs
@@ -12,12 +12,12 @@
 //! echo '{"ts": "2022-01-01 13:00:00.000000000", "temperature": 5.3}' |nc localhost 54300
 //! ```
 
-use springql_core::api::{SpringConfig, SpringPipelineHL};
+use springql_core::api::{SpringConfig, SpringPipeline};
 
 fn main() {
     const SOURCE_PORT: u16 = 54300;
 
-    let pipeline = SpringPipelineHL::new(&SpringConfig::default()).unwrap();
+    let pipeline = SpringPipeline::new(&SpringConfig::default()).unwrap();
 
     pipeline
         .command(
