+++ springql-core/src/api.rs
@@ -6,7 +6,7 @@ mod low_level_rs;
 pub use crate::{
     api::{
         error::{Result, SpringError},
-        high_level_rs::{SpringPipelineHL, SpringRowHL},
+        high_level_rs::{SpringPipeline, SpringRow},
         low_level_rs::{spring_config::*, SpringConfig},
     },
     stream_engine::{
