+++ springql-core/src/api/low_level_rs.rs
@@ -4,128 +4,6 @@
 //!
 //! C API and high-level Rust API are provided separately.
 
-mod engine_mutex;
 pub mod spring_config;
 
-use std::{sync::Once, thread, time::Duration};
-
 pub use crate::api::low_level_rs::spring_config::*;
-use crate::{
-    api::{error::Result, low_level_rs::engine_mutex::EngineMutex},
-    pipeline::name::QueueName,
-    sql_processor::SqlProcessor,
-    stream_engine::{command::Command, SinkRow},
-};
-
-fn setup_logger() {
-    static INIT: Once = Once::new();
-
-    INIT.call_once(|| {
-        let _ = env_logger::builder()
-            .is_test(false) // To enable color. Logs are not captured by test framework.
-            .try_init();
-        log_panics::init();
-    });
-
-    log::info!("setup_logger(): done");
-}
-/// Connection object.
-///
-/// 1 stream pipeline has only 1 connection.
-/// In other words, the lifecycle of SpringConnection and internal stream pipeline are the same.
-#[derive(Debug)]
-pub struct SpringPipeline {
-    engine: EngineMutex,
-    sql_processor: SqlProcessor,
-}
-
-/// Row object from an in memory queue.
-#[derive(Debug)]
-pub struct SpringRow(pub(in crate::api) SinkRow);
-
-impl From<SinkRow> for SpringRow {
-    fn from(sink_row: SinkRow) -> Self {
-        Self(sink_row)
-    }
-}
-
-/// Creates and open an in-process stream pipeline.
-pub fn spring_open(config: &SpringConfig) -> Result<SpringPipeline> {
-    setup_logger();
-
-    let engine = EngineMutex::new(config);
-    let sql_processor = SqlProcessor::default();
-
-    Ok(SpringPipeline {
-        engine,
-        sql_processor,
-    })
-}
-
-/// Execute commands (DDL).
-///
-/// # Failure
-///
-/// - [SpringError::Sql](crate::api::error::SpringError::Sql) when:
-///   - Invalid SQL syntax.
-///   - Refers to undefined objects (streams, pumps, etc)
-///   - Other semantic errors.
-/// - [SpringError::InvalidOption](crate::api::error::SpringError::Sql) when:
-///   - `OPTIONS` in `CREATE` statement includes invalid key or value.
-pub fn spring_command(pipeline: &SpringPipeline, sql: &str) -> Result<()> {
-    let mut engine = pipeline.engine.get()?;
-
-    let command = pipeline
-        .sql_processor
-        .compile(sql, engine.current_pipeline())?;
-
-    match command {
-        Command::AlterPipeline(c) => engine.alter_pipeline(c),
-    }
-}
-
-/// Pop a row from an in memory queue. This is a blocking function.
-///
-/// **Do not call this function from threads.**
-/// If you need to pop from multiple in-memory queues using threads, use `spring_pop_non_blocking()`.
-/// See: <https://github.com/SpringQL/SpringQL/issues/125>
-///
-/// # Failure
-///
-/// - [SpringError::Unavailable](crate::api::error::SpringError::Unavailable) when:
-///   - queue named `queue` does not exist.
-pub fn spring_pop(pipeline: &SpringPipeline, queue: &str) -> Result<SpringRow> {
-    const SLEEP_MSECS: u64 = 10;
-
-    let mut engine = pipeline.engine.get()?;
-
-    loop {
-        if let Some(sink_row) =
-            engine.pop_in_memory_queue_non_blocking(QueueName::new(queue.to_string()))?
-        {
-            return Ok(SpringRow::from(sink_row));
-        } else {
-            thread::sleep(Duration::from_millis(SLEEP_MSECS));
-        }
-    }
-}
-
-/// Pop a row from an in memory queue. This is a non-blocking function.
-///
-/// # Returns
-///
-/// - `Ok(Some)` when at least a row is in the queue.
-/// - `None` when no row is in the queue.
-///
-/// # Failure
-///
-/// - [SpringError::Unavailable](crate::api::error::SpringError::Unavailable) when:
-///   - queue named `queue` does not exist.
-pub fn spring_pop_non_blocking(
-    pipeline: &SpringPipeline,
-    queue: &str,
-) -> Result<Option<SpringRow>> {
-    let mut engine = pipeline.engine.get()?;
-    let sink_row = engine.pop_in_memory_queue_non_blocking(QueueName::new(queue.to_string()))?;
-    Ok(sink_row.map(SpringRow::from))
-}
