+++ springql-core/tests/test_support/mod.rs
@@ -2,12 +2,12 @@
 
 use std::time::Duration;
 
-use springql_core::{api::SpringConfig, api::SpringPipelineHL};
+use springql_core::{api::SpringConfig, api::SpringPipeline};
 use springql_foreign_service::sink::ForeignSink;
 
 #[allow(dead_code)]
-pub(crate) fn apply_ddls(ddls: &[String], config: SpringConfig) -> SpringPipelineHL {
-    let pipeline = SpringPipelineHL::new(&config).unwrap();
+pub(crate) fn apply_ddls(ddls: &[String], config: SpringConfig) -> SpringPipeline {
+    let pipeline = SpringPipeline::new(&config).unwrap();
     for ddl in ddls {
         pipeline.command(ddl).unwrap();
     }
